+++ src/main/java/com/SWP391/G3PCoffee/repository/ProductRepository.java
@@ -4,7 +4,13 @@ import com.SWP391.G3PCoffee.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.Optional;
+
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+
+
 
 }
